Update from codeil2 (#3)
* bump singleuser version

* modify space to trigger build

* bump singleuser

* add singleuser to path, activate singleuser env

* see if installing entire environment fixes gcov linking issue

* Revert "see if installing entire environment fixes gcov linking issue"

This reverts commit 6c41a1fa7d3fe34cfbed0d1e266cf02d8cb5cf9b.

* Resolve "Write Additional Tests Where Needed"

* Resolve "Feature Requests"

* Update singleuser for cdsdashboards and dask status

* Add fixes for CDSdashboards

* remove duplicate CMD option

* updates singleuser with htop and fix for matplotlib

* Resolve "Generate TDP Output"

* Add new viz libraries based on OPAL-450

* create image without pytorch deps

* add back pytorch deps

* (no-pytorch deps) OPAL-470 conf files in /home/jovyan/conf

* Add back pytorch deps

* test pulling git from pipeline

* Further testing on cloning git

* testing: correct clone command

* more git testing

* testing: correct typo

* testing: more testing

* testing: modify url

* testing: try with user set

* testing: add proper format for token cloning

* test opal-scripts copy

* Opal scripts inside tip container. Add back pytorch

* Move conf and opal directories

* remove empty /home/jovyan/conf directory that came from upstream

* Pulling latest opal code: add space to trigger build

* add gtest/gmock, fix bug where tip isn't being installed in singleuser

* OPAL not pulling latest commit: remove space to trigger build

* space to trigger build

* force remove opal directory before copying again

* run without pytorch

* add back pytorch

* test without RM command on the opal directory

* updated dockerfile to pull new opal repo

* Update dockerfile to pip install opal requirements

* Pull newest opal repo: add space

* build without pytorch

* build with pytorch

* add pip installation to pytorch env

* correct file name for pip install

* Add back tip package (previously missing)

* Resolve "Complete ARINC429 Translation"

* Change pytorch image to the new dedicated pytorch image from IB

* Resolve "Improve CLI; remove conf files"

* bump pytorch version

* remove test key that's making twistlock mad

* Resolve "CLI: Auto-adjust to terminal width"

* Resolve "e2e: execute ARINC429 translation"

* update singleuser which contains codeserver

* remove key files

* build without pytorch

* Resolve "Refactor ParseManager"

* reverting to previous version of singleuser image

* space to trigger build and grab opal repo

* bump singleuser version

* Update dockerfile for new pip packages

* Pulling latest opal code

* Resolve "Test ParseManager"

* space to trigger build

* start pipeline for update to opal repo

* adjust thing to start pipeline

* space to start pipeline for OPAL repo update

* space to trigger build

* bump singleuser version

* Resolve "Additional metadata for ARINC429"

* Resolve "Include 1553 Status Words and Structs in Parsed Data"

* Resolve "Create DTS429 example yaml"

* Resolve "Github Prep: Simplify build system"

* toggle a bit to make IB build to a new tag

* inserted misspelled image name of singleuser in dockerfile for proper build

* Resolve "Implement CLI for pqcompare and video extractor"

* removed the test.key that was triggering twistlock

* Updated singleuser ironbank image

* Update Dockerfile to revert back to singeluser_v12

* Trying v13 singleuser again

* Remove caching and codecov in ubuntu build

* Try ccache action

* Enable code coverage

* Modify path for ccache

* Fix ubuntu build yaml syntax

* Fix ubuntu build yaml syntax again

* Simplify codecov.yml; use codecov v3

* echo PATH to debug

* echo .ccache size before after build

* Remove conda save pkg as artifact step

* Print .ccache size in several steps

* Remove ccache size check before ccache step

+++ cpp/parquet_context/include/column_data.h
@@ -232,9 +232,9 @@ class ColumnData
         }
         pointer_set_ = true;
         if (is_list_)
-            initial_max_row_size_ = data.size() / list_size_;
+            initial_max_row_size_ = static_cast<int>(data.size()) / list_size_;
         else
-            initial_max_row_size_ = data.size();
+            initial_max_row_size_ = static_cast<int>(data.size());
     }
 
     void SetInputData(std::shared_ptr<InputDataBase> input_data)
